migrated fetchData code over to main, replaced fetchProjectData function
