+++ components/classroom/main.js
@@ -3,31 +3,41 @@
 import { getProjectIDFromURL } from './groups/utils/project.js';
 import Roles from "./roles.mjs";
 import hasPermission from "./hasPermission";
-import { Action, Scope, Entity } from './permissions_parameters.mjs';
+import TPEN from '../../TPEN/index.mjs';
 
 
 // Function to fetch project data from the TPEN API
 async function fetchProjectData(projectId) {
-    const bearerToken = 'eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6Ik9FVTBORFk0T1RVNVJrRXlOREl5TTBFMU1FVXdNMFUyT0RGQk9UaEZSa1JDTXpnek1FSTRNdyJ9.eyJodHRwOi8vc3RvcmUucmVydW0uaW8vYWdlbnQiOiJodHRwczovL3N0b3JlLnJlcnVtLmlvL3YxL2lkLzY2YjliMzE5OTMyNTgyMzBjYTM4NmY4NiIsImh0dHA6Ly9yZXJ1bS5pby9hcHBfZmxhZyI6WyJ0cGVuIl0sImh0dHA6Ly9kdW5iYXIucmVydW0uaW8vYXBwX2ZsYWciOlsidHBlbiJdLCJodHRwOi8vcmVydW0uaW8vdXNlcl9yb2xlcyI6eyJyb2xlcyI6WyJkdW5iYXJfdXNlcl9wdWJsaWMiLCJnbG9zc2luZ191c2VyX3B1YmxpYyIsImxyZGFfdXNlcl9wdWJsaWMiLCJyZXJ1bV91c2VyX3B1YmxpYyIsInRwZW5fdXNlcl9wdWJsaWMiXX0sImh0dHA6Ly9kdW5iYXIucmVydW0uaW8vdXNlcl9yb2xlcyI6eyJyb2xlcyI6WyJkdW5iYXJfdXNlcl9wdWJsaWMiLCJnbG9zc2luZ191c2VyX3B1YmxpYyIsImxyZGFfdXNlcl9wdWJsaWMiLCJyZXJ1bV91c2VyX3B1YmxpYyIsInRwZW5fdXNlcl9wdWJsaWMiXX0sImlzcyI6Imh0dHBzOi8vY3ViYXAuYXV0aDAuY29tLyIsInN1YiI6ImF1dGgwfDY2YjliMzE4YzZhMjU4MDZjOTgxMDg3YyIsImF1ZCI6WyJodHRwczovL2N1YmFwLmF1dGgwLmNvbS9hcGkvdjIvIiwiaHR0cHM6Ly9jdWJhcC5hdXRoMC5jb20vdXNlcmluZm8iXSwiaWF0IjoxNzI4Njc1MDczLCJleHAiOjE3Mjg2ODIyNzMsInNjb3BlIjoib3BlbmlkIHByb2ZpbGUgZW1haWwgdXBkYXRlOmN1cnJlbnRfdXNlcl9tZXRhZGF0YSBvZmZsaW5lX2FjY2VzcyIsImF6cCI6ImJCdWdGTVdIVW8xT2huU1pNcFlVWHhpM1kxVUpJN0tsIn0.TRmYqcEIGh5lvt_6mjX6CULsiIthh-P-gWLAl_JDuKjZsbbw-gSzX5ogQ70Iya9vxGdbBemYdKMqW2I8CN7w-4MCt4GMQvLNDXbaf8HUs4ohVqMzkpRvjjqwHtWzrIYO_f8ceKU8KFeu3czGmZ0_1_R30zTrEty0WMEVfhAI1v7qUiE5VY5tef-rbX19lcAMUku4b0xs8lKzlKFzsekSFtZsJ-pOoytMLWoEIeU_hxE8AmdCSm3W9Hg3jAVVIBDMKpY0foI4BSAsCWzfT9qnjBTbKtixM4k4MWyAJISERn9coIJFOG9ae6GP8JiVrIMD9je0gxDSxik8HnRINBsmLA'
-    const apiUrl = `https://dev.api.t-pen.org/project/${projectId}`;
-    
-    try {
-        const response = await fetch(apiUrl, {
+    const PROJECT_FORM = document.getElementById(projectId);
+    PROJECT_FORM.TPEN = new TPEN();
+    TPEN.attachAuthentication(PROJECT_FORM);
+
+    if (PROJECT_FORM.TPEN.activeProject?._id) {
+        projectId.value = PROJECT_FORM.TPEN.activeProject._id; //is this correct? are we getting value from id or project
+        const token = TPEN.getAuthorization();
+        // until the Project class is better defined, we'll just fetch the project data directly
+        const project = await fetch(`${PROJECT_FORM.TPEN.servicesURL}/project/${PROJECT_FORM.TPEN.activeProject._id}`, {
             method: 'GET',
             headers: {
-                'Authorization': `Bearer ${bearerToken}`
+                'Authorization': `Bearer ${token}`
             }
-        });
-
-        if (!response.ok) {
-            throw new Error('Network response was not ok');
-        }
-
-        const projectData = await response.json();
-        displayProjectData(projectData);
-    } catch (error) {
-        console.error('Error fetching project data:', error);
+        })
+            .then(res => res.ok ? res.json() : Promise.reject(res.status))
+            .then(data => msg.innerHTML = `<pre>${JSON.stringify(data, null, 2)}</pre>`)
+            .catch(err => msg.innerHTML = `<pre>${JSON.stringify(err, null, 2)}</pre>`);
     }
+
+    PROJECT_FORM.addEventListener('submit', async function (event) {
+        const projectID = projectId.value;
+        location.href = '?projectID=' + projectID;  // Redirect to the same page with the projectID in the query string
+    });
+
+    clickList.addEventListener('click', function (event) {
+        const LI = event.target.closest('LI');
+        if (LI) {
+            location.href = '?projectID=' + event.target.getAttribute('tpen-project-id');
+        }
+    });
 }
 
 // Function to display project data in the HTML
